requirements.txt
dsmr-parser
paho-mqtt